+++ termflow/ansi/__init__.py
@@ -45,6 +45,7 @@ from termflow.ansi.style import (
     LINK,
     STRIKEOUT,
     UNDERLINE,
+    make_clipboard_copy,
     make_link,
 )
 from termflow.ansi.utils import (
@@ -91,6 +92,7 @@ __all__ = [
     "hex2rgb",
     "hsv_to_rgb",
     "is_ansi_code",
+    "make_clipboard_copy",
     "make_link",
     "number_to_superscript",
     "parse_sgr_params",